#7 Clean up package structure and first version of keep alive
