+++ src/main/java/de/havemann/lukas/vanillahttp/VanillaHttpServer.java
@@ -1,0 +1,19 @@
+package de.havemann.lukas.vanillahttp;
+
+import org.springframework.boot.Banner;
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+
+
+@SpringBootApplication
+public class VanillaHttpServer {
+
+    public static void main(String[] args) {
+        new SpringApplicationBuilder(VanillaHttpServer.class)
+                .web(WebApplicationType.NONE)
+                .bannerMode(Banner.Mode.CONSOLE)
+                .run(args);
+    }
+}
+
