+++ src/main/java/de/havemann/lukas/vanillahttp/dispatcher/ClientConnectionDispatcher.java
@@ -1,0 +1,8 @@
+package de.havemann.lukas.vanillahttp.dispatcher;
+
+import java.net.Socket;
+
+public interface ClientConnectionDispatcher {
+
+    boolean dispatch(Socket clientSocket);
+}
