+++ src/main/java/de/havemann/lukas/vanillahttp/protocol/specification/HttpMethod.java
@@ -1,0 +1,8 @@
+package de.havemann.lukas.vanillahttp.protocol.specification;
+
+/**
+ * Supported HTTP-Methods
+ */
+public enum HttpMethod implements ProtocolRepresentation {
+    GET, HEAD
+}
