--- src/main/java/de/havemann/lukas/vanillahttpserver/protocol/specification/HttpHeaderField.java
@@ -1,29 +1,0 @@
-package de.havemann.lukas.vanillahttpserver.protocol.specification;
-
-/**
- * list of http header fields that are explicitly handled
- */
-public enum HttpHeaderField implements ProtocolRepresentation {
-
-    CONNECTION("Connection"),
-    CONTENT_TYPE("ContentType"),
-    TRANSFER_ENCODING("Transfer-Encoding"),
-
-    E_TAG("ETag"),
-    IF_MATCH("If-Match"),
-    IF_NONE_MATCH("If-None-Match"),
-    IF_MODIFIED_SINCE("If-Modified-Since");
-
-    public static final String KEY_VALUE_DELIMITER = ": ";
-
-    private final String headerFieldName;
-
-    HttpHeaderField(String headerFieldName) {
-        this.headerFieldName = headerFieldName;
-    }
-
-    @Override
-    public String getRepresentation() {
-        return headerFieldName;
-    }
-}
