--- src/main/java/de/havemann/lukas/vanillahttpserver/protocol/specification/HttpMethod.java
@@ -1,8 +1,0 @@
-package de.havemann.lukas.vanillahttpserver.protocol.specification;
-
-/**
- * Supported HTTP-Methods
- */
-public enum HttpMethod implements ProtocolRepresentation {
-    GET, HEAD
-}
