--- src/main/java/de/havemann/lukas/vanillahttpserver/service/ClientConnectionHandler.java
@@ -1,8 +1,0 @@
-package de.havemann.lukas.vanillahttpserver.service;
-
-import java.net.Socket;
-
-public interface ClientConnectionHandler {
-
-    boolean offer(Socket clientSocket);
-}
